Remove TestWorkQuality, and move TestSiteConfigurationHasChanged into test_listeners.py
